openshift: fix crash
Regression caused by pydantic v2 migration.

```
1 validation error for Uptime
seconds
Input should be a valid integer, got a number with a fractional part [type=int_from_float, input_value=4559.20867061615, input_type=float]
```

CMK-15829

+++ cmk/special_agents/utils/node_exporter.py
@@ -103,7 +103,7 @@ class CPULoad(pydantic.BaseModel):
 class Uptime(pydantic.BaseModel):
     """section: prometheus_uptime_v1"""
 
-    seconds: int
+    seconds: float
 
 
 class FilesystemInfo:
